pyproject.toml
[tool.poetry]
name = "MobiT"
version = "0.1.0"
description = ""
authors = ["sophiesg <sjiahui27@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.77.0"
pyyaml = "^6.0.2"
opencv-python = "^4.11.0.86"
pyshine = "^0.0.9"
colorama = "^0.4.6"
click = "^8.2.0"
pillow = "^11.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
